Improve implementation for firsthit.
This implementation passes all tests. There is a subtle difference to
the previous implementation.

Previously, the program iterated through all segments in question, and
filtered those that did not intersect with the ray at all. Deciding for
intersection between a segment and a ray is computationally slower than
between two segments. Hence, the new implementation skips this first
step and directly moves to finding the nearest intersection as was
already done in a later step in the previous algorithm.